Перенести в SharedPreferencesClient логику управления историей поиска треков
